Added new unchecked exception for handling I/O exceptions. Used it in DAOs. Moved exceptions to a new package

+++ src/main/java/pl/edu/wit/studentManagement/service/StudentService.java
@@ -3,7 +3,7 @@ package pl.edu.wit.studentManagement.service;
 import pl.edu.wit.studentManagement.service.dto.student.CreateStudentDto;
 import pl.edu.wit.studentManagement.service.dto.student.StudentDto;
 import pl.edu.wit.studentManagement.service.dto.student.UpdateStudentDto;
-import pl.edu.wit.studentManagement.validation.ValidationException;
+import pl.edu.wit.studentManagement.exceptions.ValidationException;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -20,6 +20,7 @@ import java.util.stream.Collectors;
 public class StudentService {
     private final Dao<Student> studentDao;
     private final Dao<StudentGroup> studentGroupDao;
+    private final Dao<Grade> gradeDao;
 
     /**
      * Constructs a StudentService with the specified data access object.
@@ -27,9 +28,10 @@ public class StudentService {
      * @param studentDao the DAO used for student persistence
      * @param studentGroupDao the DAO used for student group persistence
      */
-    StudentService(Dao<Student> studentDao, Dao<StudentGroup> studentGroupDao) {
+    StudentService(Dao<Student> studentDao, Dao<StudentGroup> studentGroupDao, Dao<Grade> gradeDao) {
         this.studentDao = studentDao;
         this.studentGroupDao = studentGroupDao;
+        this.gradeDao = gradeDao;
     }
 
     /**
@@ -92,16 +94,27 @@ public class StudentService {
         return StudentMapper.toDto(student);
     }
 
+
     /**
-     * Deletes a student by ID.
+     * Deletes a student by their unique identifier.
+     * <p>
+     * The student can only be deleted if they have no associated grades.
      *
-     * @param id the ID of the student to delete
-     * @return true if deletion was successful, false otherwise
+     * @param id the unique identifier of the student to delete
+     * @return {@code true} if the student was successfully deleted, {@code false} if the student did not exist
+     * @throws ValidationException if the student has existing grades and cannot be deleted
      */
-    public boolean deleteStudent(UUID id) {
-        // TODO: add validation like, if student has grades or is assigned to a group
+    public boolean deleteStudent(UUID id) throws ValidationException {
+        var studentGrades = gradeDao.getAll()
+                .stream()
+                .filter(g -> g.getStudentId().equals(id))
+                .collect(Collectors.toList());
+
+        if (!studentGrades.isEmpty()) throw new ValidationException("student.hasGrades");
+
         return studentDao.delete(id);
     }
+
 
     /**
      Searches for students matching the specified query.
